requirements.txt
asgiref==3.7.2
Django==4.2.3
sqlparse==0.4.4
typing_extensions==4.7.1
psycopg2-binary==2.9.6
python-dotenv==0.19.2
djangorestframework==3.14.0
django-filter==23.2
